fix(sceneview): address 5-agent review of haptic phase 1 (#1901)
Fixes the 2 BLOCKING + 7 MAJOR + MINOR findings from the independent
5-agent review of PR #1921:

BLOCKING
- API 24-25 crash: the legacy preset fallbacks passed
  VibrationEffect.DEFAULT_AMPLITUDE — VibrationEffect is API 26, so
  referencing its static field on a minSdk-24 device throws
  NoClassDefFoundError. Introduce a package-internal HAPTIC_DEFAULT_AMPLITUDE
  sentinel; only SystemHapticEngine's >= O branch resolves the real
  VibrationEffect symbol. The < O path now names no VibrationEffect member.
- Web continuous() was missing — add continuous(intensity, durationMs)
  mapping to navigator.vibrate(durationMs); intensity documented as ignored.

MAJOR
- pattern() of all-zero timings passed the isEmpty() guard and made
  createWaveform throw IllegalArgumentException — now a no-op.
- continuous() was not cancellable — add SceneViewHaptic.cancel() and a
  DisposableEffect in rememberHapticFeedback() that cancels on dispose.
- iOS engine died after backgrounding — play() now calls engine.start()
  (cheap/idempotent) before makePlayer; stoppedHandler comment corrected.
- iOS continuous() unit divergence — changed from TimeInterval seconds to
  durationMs: Int milliseconds, matching Android and Web.
- HapticEvent field-type divergence — Android durationMs/delayMs aligned to
  Int (Long widening kept at the single Vibrator boundary).
- SystemHapticEngine SDK branches were untested — add Robolectric
  SystemHapticEngineRobolectricTest parameterised by SDK (24/25/27/33),
  pinning the legacy-overload path and the createOneShot/createPredefined
  fallbacks. Add Robolectric to :sceneview test deps.
- Haptic API added to agents/sceneview and agents/sceneview-web skills.

MINOR
- rememberHapticFeedback() keys remember on applicationContext, not
  LocalContext.
- Removed misleading @RequiresApi(O) on the amplitude-waveform path.
- Added tests: negative-intensity clamp to amplitude 1, negative
  pattern timings coerced to 0, degraded-path no-op + single Log.d
  (ShadowLog).

llms.txt + both mirrors and the changelog fragment updated for the new
continuous() signature and cancel().

+++ sceneview/src/main/java/io/github/sceneview/haptic/HapticEvent.kt
@@ -22,17 +22,22 @@ package io.github.sceneview.haptic
  * - **Web** — events render via `navigator.vibrate(durationsMs[])`; [intensity] and
  *   [sharpness] are ignored (the Web Vibration API takes durations only).
  *
+ * The field types mirror the iOS `HapticEvent` (`Int` milliseconds) so the
+ * same pattern definitions are byte-for-byte portable across platforms.
+ *
  * @property intensity Strength of the haptic, `0.0` (off) .. `1.0` (max). Values
  * outside the range are clamped.
  * @property sharpness Wave-shape hint, `0.0` (round/dull) .. `1.0` (sharp/clicky).
  * Honoured on iOS; ignored on Android / Web in this release.
- * @property durationMs How long this event vibrates, in milliseconds.
+ * @property durationMs How long this event vibrates, in milliseconds. Negative
+ * values are coerced to `0`.
  * @property delayMs Delay before this event starts, relative to the end of the
- * previous event (or to `pattern()` invocation for the first event).
+ * previous event (or to `pattern()` invocation for the first event). Negative
+ * values are coerced to `0`.
  */
 public data class HapticEvent(
     val intensity: Float,
     val sharpness: Float,
-    val durationMs: Long,
-    val delayMs: Long = 0,
+    val durationMs: Int,
+    val delayMs: Int = 0,
 )
